requirements.txt
# Server
python-dateutil
schedule

# Client
paramiko

# Both
websockets